:recycle: Standardize use general ledger coding pattern.
