+++ web/src/modules/travel-authorizations/components/edit-my-travel-authorization-expense-page/RequestApprovalForm.vue
@@ -71,7 +71,7 @@ import { required } from "@/utils/validators"
 
 import { useSnack } from "@/plugins/snack-plugin"
 import { useTravelAuthorization } from "@/use/use-travel-authorization"
-import { useGeneralLedgerCodings } from "@/use/general-ledger-codings"
+import { useGeneralLedgerCodings } from "@/use/use-general-ledger-codings"
 import { useExpenses, TYPES, EXPENSE_TYPES } from "@/use/use-expenses"
 
 import SearchableUserEmailCombobox from "@/components/SearchableUserEmailCombobox"
@@ -97,11 +97,18 @@ const {
   fetch: fetchTravelAuthorization,
   expenseClaim,
 } = useTravelAuthorization(props.travelAuthorizationId)
+
+const generalLedgerCodingOptions = computed(() => ({
+  where: {
+    travelAuthorizationId: props.travelAuthorizationId,
+  },
+}))
 const {
   generalLedgerCodings,
   isLoading: isLoadingGeneralLedgerCodings,
   fetch: fetchGeneralLedgerCodings,
-} = useGeneralLedgerCodings()
+} = useGeneralLedgerCodings(generalLedgerCodingOptions)
+
 const expenseOptions = computed(() => ({
   where: {
     travelAuthorizationId: props.travelAuthorizationId,
@@ -134,11 +141,7 @@ onMounted(async () => {
 async function refresh() {
   await Promise.all([
     await fetchTravelAuthorization(),
-    await fetchGeneralLedgerCodings({
-      where: {
-        travelAuthorizationId: props.travelAuthorizationId,
-      },
-    }),
+    await fetchGeneralLedgerCodings(),
     await fetchExpenses(),
   ])
 }
